// src/modules/graphql/arweave.net/types/Schema.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
};

export type Query = {
  __typename?: "Query";
  transaction?: Maybe<Transaction>;
  transactions: TransactionConnection;
  block?: Maybe<Block>;
  blocks: BlockConnection;
};

export type QueryTransactionArgs = {
  id: Scalars["ID"]["input"];
};

export type QueryTransactionsArgs = {
  ids?: InputMaybe<Array<Scalars["ID"]["input"]>>;
  owners?: InputMaybe<Array<Scalars["String"]["input"]>>;
  recipients?: InputMaybe<Array<Scalars["String"]["input"]>>;
  tags?: InputMaybe<Array<TagFilter>>;
  bundledIn?: InputMaybe<Array<Scalars["ID"]["input"]>>;
  block?: InputMaybe<BlockFilter>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
  after?: InputMaybe<Scalars["String"]["input"]>;
  sort?: InputMaybe<SortOrder>;
};

export type QueryBlockArgs = {
  id?: InputMaybe<Scalars["String"]["input"]>;
};

export type QueryBlocksArgs = {
  ids?: InputMaybe<Array<Scalars["ID"]["input"]>>;
  height?: InputMaybe<BlockFilter>;
  first?: InputMaybe<Scalars["Int"]["input"]>;
  after?: InputMaybe<Scalars["String"]["input"]>;
  sort?: InputMaybe<SortOrder>;
};

export enum SortOrder {
  HeightAsc = "HEIGHT_ASC",
  HeightDesc = "HEIGHT_DESC",
}

export type TagFilter = {
  name: Scalars["String"]["input"];
  values: Array<Scalars["String"]["input"]>;
  op?: InputMaybe<TagOperator>;
};

export type BlockFilter = {
  min?: InputMaybe<Scalars["Int"]["input"]>;
  max?: InputMaybe<Scalars["Int"]["input"]>;
};
// ...
export enum TagOperator {
  Eq = "EQ",
  Neq = "NEQ",
}